distance to 200 km

+++ src/main/java/com/xecoder/service/service/MessagesService.java
@@ -85,7 +85,7 @@ public class MessagesService extends AbstractService<Messages> {
         query.skip(calcSkipNum(page, size)).limit(size);
         List<Messages> list = new ArrayList<>();
         //按经纬度搜索
-            NearQuery nq = NearQuery.near(point.getX(), point.getY(), Metrics.KILOMETERS).maxDistance(new Double(20)).query(query);//单位: 20千米
+            NearQuery nq = NearQuery.near(point.getX(), point.getY(), Metrics.KILOMETERS).maxDistance(new Double(200)).query(query);//单位: 20千米
             GeoResults<Messages> empGeoResults = mongoTemplate.geoNear(nq, Messages.class);
             if (empGeoResults != null) {
                 for (GeoResult<Messages> e : empGeoResults) {
